# YJIT: Inline basic Ruby methods