@maximecb 2023-01-11
I think this is an interesting experiment, but I'm inclined to not want to merge this since it adds more complexity and special-casing to our method call logic and other things, but produces no visible performance gains. I would rather that if we go for inlining Ruby methods, we try to tackle it in a way that is more general and can inline more methods. This might require re-architecting our context object, which I would also prefer we discuss beforehand.

At the moment, based on the stats you've previously gathered, it also seems like C methods dwarf Ruby methods in terms of call frequency, so maybe we should focus our effort there first. However, those stats don't tell us how the overhead of C method calls compares to the overhead of Ruby method calls in terms of CPU instructions per call. It would be interesting to try and estimate the respective overhead of each.

All of that being said, I'm curious if we'd see a performance win if we could just inline those methods that all they do is just call `Hash#[]` and then return... 🤔