@k0kubun 2023-01-11
> I think this is an interesting experiment, but I'm inclined to not want to merge this since it adds more complexity and special-casing to our method call logic and other things, but produces no visible performance gains. I would rather that if we go for inlining Ruby methods, we try to tackle it in a way that is more general and can inline more methods. This might require re-architecting our context object, which I would also prefer we discuss beforehand.

Yeah, I agree. It was an interesting finding that this doesn't make a significant difference, but it's not really useful other than finding that. We could always copy code from here if we want to reuse this. When we merge something, let's start from something that makes an actual impact. 

> You have a couple of good asm comments in here that don't look related to inlining. So even if inlining doesn't get merged, it's probably worth making sure the comments still get merged.

I'll file a separate PR for it :+1: