@k0kubun 2023-01-11
> All of that being said, I'm curious if we'd see a performance win if we could just inline those methods that all they do is just call Hash#[] and then return... 🤔

We'd need to support side exits from inlined methods, which is what this design is deliberately scoping out, and `opt_aref` would also need to have some optimization that uses chain guards per key. But I agree that it's one promising use case for inlining, given the stats we have and what we've discussed in https://github.com/Shopify/ruby/issues/482.